requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.5.2
pydantic-settings==2.1.0

# Database
supabase==2.3.0
redis==5.0.1
aioredis==2.0.1

# AI Services
google-generativeai==0.3.2
elevenlabs==0.2.27
deepgram-sdk==3.1.0
pyannote.audio==3.1.1
torch==2.1.1
torchaudio==2.1.1

# Audio Processing
soundfile==0.12.1
numpy==1.26.2
scipy==1.11.4
webrtcvad==2.0.10

# HTTP Client
httpx==0.25.2
backoff==2.2.1

# Monitoring & Logging
prometheus-client==0.19.0
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.14

# SignalWire
signalwire==2.0.0

# Core Dependencies
gunicorn>=21.2.0,<22.0.0
python-dotenv>=0.19.0,<0.20.0
httpx>=0.24.0,<0.25.0
redis>=4.0.0,<5.0.0
supabase>=0.0.2,<0.1.0

# AI/ML Dependencies
google-generativeai>=0.3.0,<0.4.0
elevenlabs>=0.2.0,<0.3.0
deepgram-sdk>=2.0.0,<3.0.0
pyannote.audio>=2.0.0,<3.0.0
torch>=1.9.0,<2.0.0
torchaudio>=0.9.0,<1.0.0

# Audio Processing
soundfile>=0.10.0,<0.11.0
numpy>=1.21.0,<1.22.0
scipy>=1.7.0,<1.8.0

# Testing and Development
pytest>=6.2.0,<6.3.0
pytest-asyncio>=0.15.0,<0.16.0
black>=21.7b0,<21.8b0
isort>=5.9.0,<5.10.0
flake8>=3.9.0,<3.10.0
mypy>=0.910,<0.911

# Documentation
mkdocs>=1.2.0,<1.3.0
mkdocs-material>=7.0.0,<7.1.0

# Security
python-jose[cryptography]>=3.3.0,<3.4.0
passlib[bcrypt]>=1.7.4,<1.8.0
python-multipart>=0.0.5,<0.0.6

# Monitoring and Logging
prometheus-client>=0.11.0,<0.12.0 